- Refactorizada la clase para generar datos aleatorios. - Ya se pueden generar cuentas, epigrafes y grupos de epígrafes aleatorios. - Ya no se muestra el botón de borrar en los listados si no hay resultados.
