+++ Core/View/Randomizer.html
@@ -5,7 +5,7 @@
     <div class="row">
         <div class="col-12">
             <h1>
-                {{ i18n.trans('generate-test-data') }}
+                <i class="fa fa-magic" aria-hidden="true"></i> {{ i18n.trans('generate-test-data') }}
                 <a href="{{ fsc.url() }}" class="btn btn-sm btn-secondary">
                     <i class="fa fa-refresh" aria-hidden="true"></i>
                 </a>
@@ -18,46 +18,109 @@
             <a href="{{ fsc.url() }}&gen=fabricantes" class="btn btn-block btn-info">
                 <i class="fa fa-columns" aria-hidden="true"></i>&nbsp; {{ i18n.trans('manufacturers') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=familias" class="btn btn-block btn-info">
                 <i class="fa fa-object-group" aria-hidden="true"></i>&nbsp; {{ i18n.trans('families') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=articulos" class="btn btn-block btn-info">
                 <i class="fa fa-cubes" aria-hidden="true"></i>&nbsp; {{ i18n.trans('products') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=articulosprov" class="btn btn-block btn-info">
                 <i class="fa fa-cubes" aria-hidden="true"></i>&nbsp; {{ i18n.trans('supplier-products') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=agentes" class="btn btn-block btn-info">
                 <i class="fa fa-users" aria-hidden="true"></i>&nbsp; {{ i18n.trans('agents') }}
             </a>
+            <br/>
         </div>
+    </div>
+    <div class="row">
+        <div class="col-12">
+            <h3>{{ i18n.trans('suppliers') }}</h3>
+            <hr/>
+        </div>
+    </div>
+    <div class="row">
         <div class="col-4">
             <a href="{{ fsc.url() }}&gen=proveedores" class="btn btn-block btn-info">
                 <i class="fa fa-users" aria-hidden="true"></i>&nbsp; {{ i18n.trans('suppliers') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=pedidosprov" class="btn btn-block btn-info">
                 <i class="fa fa-list" aria-hidden="true"></i>&nbsp; {{ i18n.trans('supplier-orders') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=albaranesprov" class="btn btn-block btn-info">
                 <i class="fa fa-list" aria-hidden="true"></i>&nbsp; {{ i18n.trans('supplier-albaranes') }}
             </a>
+            <br/>
         </div>
+    </div>
+    <div class="row">
+        <div class="col-12">
+            <h3>{{ i18n.trans('customers') }}</h3>
+            <hr/>
+        </div>
+    </div>
+    <div class="row">
         <div class="col-4">
             <a href="{{ fsc.url() }}&gen=grupos" class="btn btn-block btn-info">
                 <i class="fa fa-folder" aria-hidden="true"></i>&nbsp; {{ i18n.trans('customer-groups') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=clientes" class="btn btn-block btn-info">
                 <i class="fa fa-users" aria-hidden="true"></i>&nbsp; {{ i18n.trans('customers') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=presupuestoscli" class="btn btn-block btn-info">
                 <i class="fa fa-list" aria-hidden="true"></i>&nbsp; {{ i18n.trans('customer-presupuestos') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=pedidoscli" class="btn btn-block btn-info">
                 <i class="fa fa-list" aria-hidden="true"></i>&nbsp; {{ i18n.trans('customer-orders') }}
             </a>
+            <br/>
+        </div>
+        <div class="col-4">
             <a href="{{ fsc.url() }}&gen=albaranescli" class="btn btn-block btn-info">
                 <i class="fa fa-list" aria-hidden="true"></i>&nbsp; {{ i18n.trans('customer-albaranes') }}
             </a>
+            <br/>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-12">
+            <h3>{{ i18n.trans('accounting') }}</h3>
+            <hr/>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-4">
+            <a href="{{ fsc.url() }}&gen=cuentas" class="btn btn-block btn-info">
+                <i class="fa fa-book" aria-hidden="true"></i>&nbsp; {{ i18n.trans('accounting-accounts') }}
+            </a>
+            <br/>
         </div>
     </div>
 </div>
